requirements.txt
python-telegram-bot[all]==20.7
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pyyaml==6.0.1
python-dotenv==1.0.0
aiohttp==3.9.1
APScheduler==3.10.4
